backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
gunicorn==23.0.0
numpy==2.3.2
pandas==2.3.2